[Armazenar lousa no banco] Salvando os comentários no banco de dados

+++ src/main/java/br/usp/suricato/models/Comentario.java
@@ -16,10 +16,10 @@ public class Comentario {
 	private String conteudo;
 
 	@NotNull
-	private Long posicaoHorizontal;
+	private Double posicaoHorizontal;
 
 	@NotNull
-	private Long posicaoVertical;
+	private Double posicaoVertical;
 
 	public Integer getId() {
 		return id;
@@ -37,19 +37,19 @@ public class Comentario {
 		this.conteudo = conteudo;
 	}
 
-	public Long getPosicaoHorizontal() {
+	public Double getPosicaoHorizontal() {
 		return posicaoHorizontal;
 	}
 
-	public void setPosicaoHorizontal(Long posicaoHorizontal) {
+	public void setPosicaoHorizontal(Double posicaoHorizontal) {
 		this.posicaoHorizontal = posicaoHorizontal;
 	}
 
-	public Long getPosicaoVertical() {
+	public Double getPosicaoVertical() {
 		return posicaoVertical;
 	}
 
-	public void setPosicaoVertical(Long posicaoVertical) {
+	public void setPosicaoVertical(Double posicaoVertical) {
 		this.posicaoVertical = posicaoVertical;
 	}
